+++ src/pages/functions/settings/pages/address/index.tsx
@@ -1,45 +1,23 @@
 import { Button, Modal, ModalBody, ModalContent, ModalFooter, useDisclosure } from '@heroui/react';
 import { useRef, useState } from 'react';
-import CopyToClipboard from 'react-copy-to-clipboard';
 
 import Icon from '~components/icon';
 import { FusePage } from '~components/layouts/page';
 import { FusePageTransition } from '~components/layouts/transition';
-import { Tooltip, TooltipContent, TooltipTrigger } from '~components/ui/tooltip';
 import { useCurrentState } from '~hooks/memo/current_state';
 import { useGoto } from '~hooks/memo/goto';
 import { useMarkedAddresses, useRecentAddresses } from '~hooks/store/local-secure';
-import { useSonnerToast } from '~hooks/toast';
 import { truncate_principal, truncate_text } from '~lib/utils/text';
 import type { ChainAddress, MarkedAddress } from '~types/address';
 
 import { FunctionHeader } from '../../../components/header';
 import { AddAddressDrawer, EditAddressDrawer } from './components/drawer';
-
-export const AddressMenuTooltip = ({
-    container,
-    trigger,
-    content,
-}: {
-    container?: HTMLElement | null | undefined;
-    trigger: React.ReactNode;
-    content: React.ReactNode;
-}) => {
-    return (
-        <Tooltip>
-            <TooltipTrigger>{trigger}</TooltipTrigger>
-            <TooltipContent container={container} side="bottom" align="center">
-                {content}
-            </TooltipContent>
-        </Tooltip>
-    );
-};
+import { AddressMenuTooltip } from './components/tooltip';
 
 function FunctionSettingsAddressesPage() {
     const current_state = useCurrentState();
 
     const { setHide, goto } = useGoto();
-    const toast = useSonnerToast();
 
     const [markedAddresses, { pushOrUpdateMarkedAddress, removeMarkedAddress }] = useMarkedAddresses();
     const [recentAddresses] = useRecentAddresses();
@@ -99,59 +77,10 @@ function FunctionSettingsAddressesPage() {
                                             </div>
                                             <AddressMenuTooltip
                                                 container={ref.current ?? undefined}
-                                                trigger={
-                                                    <div className="flex h-8 w-8 items-center justify-center rounded-lg duration-300 hover:bg-[#2B2B2B]">
-                                                        <span className="h-[3px] w-[3px] rounded-full bg-[#999999]"></span>
-                                                        <span className="mx-[2px] h-[3px] w-[3px] rounded-full bg-[#999999]"></span>
-                                                        <span className="h-[3px] w-[3px] rounded-full bg-[#999999]"></span>
-                                                    </div>
-                                                }
-                                                content={
-                                                    <div className="w-[120px] rounded-xl">
-                                                        <CopyToClipboard
-                                                            text={item.address.address}
-                                                            onCopy={() => {
-                                                                toast.success('Copied');
-                                                            }}
-                                                        >
-                                                            <div className="flex items-center rounded-lg p-2 duration-300 hover:bg-[#333333]">
-                                                                <Icon
-                                                                    name="icon-copy"
-                                                                    className="mr-2 h-3 w-3 shrink-0 cursor-pointer text-[#999999]"
-                                                                />
-                                                                <span>Copy</span>
-                                                            </div>
-                                                        </CopyToClipboard>
-                                                        <div
-                                                            className="flex items-center rounded-lg p-2 duration-300 hover:bg-[#333333]"
-                                                            onClick={(e) => {
-                                                                e.stopPropagation();
-                                                                setEditAddress(item);
-                                                            }}
-                                                        >
-                                                            <Icon
-                                                                name="icon-edit"
-                                                                className="mr-2 h-4 w-4 shrink-0 cursor-pointer text-[#999999]"
-                                                            />
-                                                            <span>Edit</span>
-                                                        </div>
-                                                        <div
-                                                            className="flex items-center rounded-lg p-2 duration-300 hover:bg-[#333333]"
-                                                            onClick={(e) => {
-                                                                e.stopPropagation();
-                                                                onOpen();
-                                                                setRemoveAddress(item.address);
-                                                                // removeMarkedAddress(item.address);
-                                                            }}
-                                                        >
-                                                            <Icon
-                                                                name="icon-delete"
-                                                                className="mr-2 h-4 w-4 shrink-0 cursor-pointer text-[#999999]"
-                                                            />
-                                                            <span>Delete</span>
-                                                        </div>
-                                                    </div>
-                                                }
+                                                item={item}
+                                                setEditAddress={setEditAddress}
+                                                onOpen={onOpen}
+                                                setRemoveAddress={setRemoveAddress}
                                             />
                                         </div>
                                     </div>
